% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countryAggregates.r
\name{countryAggregates}
\alias{countryAggregates}
\title{countryAggregates}
\usage{
countryAggregates(PlayerName, PlayerCountry)
}
\arguments{
\item{PlayerName}{Player name to be searched for in the cricinfo database}
}
\description{
Creates a dataframe giving player averages against countries played against
}
\examples{
Aggregates <- countryAggregates(PlayerName="Alastair Cook")
}

