% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPlayerInfo.r
\name{getPlayerInfo}
\alias{getPlayerInfo}
\title{getPlayerInfo}
\usage{
getPlayerInfo(PlayerSearch, PlayerNumber = "", BattingBowling = "Batting",
  ViewType = "Ave", Clean = "Y")
}
\arguments{
\item{PlayerNumber}{Cricinfo player number}

\item{BattingBowling}{Batting or bowling figures. Default is batting}

\item{ViewType}{Data to be returned. Default is average}

\item{Clean}{Do you want the data to be cleaned? Default is yes}

\item{view}{Do you want to show in a table? Defaults to TRUE.}
}
\description{
Scrapes player info from cricinfo
}
\examples{
playerInfo <- getPlayerInfo(PlayerNumber="11728", BattingBowling = "Batting",
					ViewType = "Inn", Clean="Y")
}

