% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isBatsmanInForm.r
\name{isBatsmanInForm}
\alias{isBatsmanInForm}
\title{isBatsmanInForm}
\usage{
isBatsmanInForm(Runs, Popsize = 10, PValue = 0.05)
}
\arguments{
\item{Runs}{Vector containing the runs for the player}

\item{Popsize}{Number of innings to consider}

\item{PValue}{P Value to compare against}
}
\description{
Compares the lastest of a batsmans innings against their career average
}
\examples{
playerInfo <- getPlayerInfo(PlayerNumber="11728", BattingBowling = "Batting", ViewType = "Inn", Clean="Y")
isBatsmanInForm(playerInfo$Runs, 0.05)
}

