% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lifetimeBattingAverage.r
\name{lifetimeBattingAverage}
\alias{lifetimeBattingAverage}
\title{lifetimeBattingAverage}
\usage{
lifetimeBattingAverage(NOFlag, runs)
}
\arguments{
\item{NOFlag}{Flag showing if the innings was not out. By default 1 is not out}

\item{runs}{Runs scored in innings}

\item{BattingBowling}{Batting or bowling figures. Default is batting}
}
\description{
Creates a dataframe giving a career batting average
}
\examples{
playerInfo <- getPlayerInfo(PlayerNumber="11728", BattingBowling = "Batting", ViewType = "Inn", Clean="Y")
playerInfo$Average <- lifetimeBattingAverage(NOFlag=playerInfo$NOFlag, runs = playerInfo$Runs)
}

