% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outMethods.r
\name{outMethods}
\alias{outMethods}
\title{outMethods}
\usage{
outMethods(PlayerName, PlayerCountry)
}
\arguments{
\item{PlayerName}{Player name to be searched for in the cricinfo database}

\item{PlayerCountry}{Country for associated player}
}
\description{
Creates a dataframe giving player outs
}
\examples{
Aggregates <- outMethods(PlayerName="Alastair Cook", PlayerCountry = "England")
}

