% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareTwoPlayerAves.r
\name{compareTwoPlayerAves}
\alias{compareTwoPlayerAves}
\title{compareTwoPlayerAves}
\usage{
compareTwoPlayerAves(Player1Name, Player1Country, Player2Name, Player2Country)
}
\arguments{
\item{Player1Country}{Country of the first player}

\item{Player2Country}{Country of the second player}

\item{Player1}{Name of the first player}

\item{Player2}{Name of the second player}
}
\description{
Uses the getPlayerInfo function to return a dataframe with the two player averages
}
\examples{
compareTwoPlayerAves("Stuart Broad", "England", "Alastair Cook", "England")
}

