% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetPlayerInfo.r
\name{getPlayerInfo}
\alias{getPlayerInfo}
\title{getPlayerInfo}
\usage{
getPlayerInfo(PlayerNumber, BattingBowling = "Batting", ViewType = "Ave",
  Clean = "Y")
}
\arguments{
\item{view}{Do you want to show in a table? Defaults to TRUE.}
}
\description{
Scrapes player info from cricinfo
}
\examples{
playerInfo <- getPlayerInfo(PlayerNumber="11728", BattingBowling = "Batting", ViewType = "Inn", Clean="Y")
}
\keyword{Batting}
\keyword{BattingBowling}
\keyword{Clean}
\keyword{Cricinfo}
\keyword{Data}
\keyword{Default}
\keyword{Do}
\keyword{PlayerNumber}
\keyword{ViewType}
\keyword{average}
\keyword{batting}
\keyword{be}
\keyword{bowling}
\keyword{cleaned?}
\keyword{data}
\keyword{figures.}
\keyword{is}
\keyword{number}
\keyword{or}
\keyword{player}
\keyword{returned.}
\keyword{the}
\keyword{to}
\keyword{want}
\keyword{yes}
\keyword{you}

