% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPlayerNumber.r
\name{getPlayerNumber}
\alias{getPlayerNumber}
\title{Return Cricinfo player number}
\usage{
getPlayerNumber(Name)
}
\arguments{
\item{Name}{Player name}
}
\description{
Returns the cricinfo player number
}
\examples{
getPlayerNumber(Name="Stuart Broad")
}

